+++ CHANGELOG.md
@@ -45,4 +45,5 @@
 
 ### Added 24/8/2026 
 
-- Clase jugador con dinero, experiencia, nivel, y el calculo para la progresión 
+- Clase jugador con dinero, experiencia, nivel, y el calculo para la progresión
+- HUD en la ciudad indicando dinero y nivel